新建单个弹窗

+++ demo/src/main/java/com/web/joke/dao/JokeUserDao.java
@@ -14,7 +14,7 @@ import java.util.List;
  * Created by roc on 2016/4/20.
  */
 @Repository
-public interface JokeUserDao extends CrudRepository<JokeUser, Integer> {
+public interface JokeUserDao extends CrudRepository<JokeUser, Long> {
 
     @Query("SELECT user FROM JokeUser user WHERE user.mail = :mail")
     List<JokeUser> getUserByMail(@Param("mail") String mail);
